♻️ Refactor: payment.ts 파일 이름 payment -> receipt로 변경
