+++ src/types/receipt.ts
